api/requirements_api.txt
fastapi==0.109.0
uvicorn==0.23.0
python-dotenv==1.0.1
mysql-connector-python==9.5.0
loguru==0.7.0
requests==2.32.0
